# assets/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FigureHook",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/webhooks/": {
            "get": {
                "tags": [
                    "discord",
                    "webhook",
                    "internal-service"
                ],
                "summary": "Get Webhooks",
                "operationId": "get_webhooks_api_v1_webhooks__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get Webhooks Api V1 Webhooks  Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DecryptedWebhookInDB"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/api/v1/webhooks/{webhook_id}": {
            "get": {
                "tags": [
                    "discord",
                    "webhook",
                    "internal-service"
                ],
                "summary": "Get Webhook",
                "operationId": "get_webhook_api_v1_webhooks__webhook_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Webhook Id",
                            "type": "string"
                        },
                        "name": "webhook_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DecryptedWebhookInDB"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "discord",
                    "webhook",
                    "internal-service"
                ],
                "summary": "Update Webhook",
                "operationId": "update_webhook_api_v1_webhooks__webhook_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Webhook Id",
                            "type": "string"
                        },
                        "name": "webhook_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebhookCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DecryptedWebhookInDB"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "discord",
                    "webhook",
                    "internal-service"
                ],
                "summary": "Delete Webhook",
                "operationId": "delete_webhook_api_v1_webhooks__webhook_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Webhook Id",
                            "type": "string"
                        },
                        "name": "webhook_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "discord",
                    "webhook",
                    "internal-service"
                ],
                "summary": "Patch Webhook",
                "operationId": "patch_webhook_api_v1_webhooks__webhook_id__patch",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Webhook Id",
                            "type": "string"
                        },
                        "name": "webhook_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebhookUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DecryptedWebhookInDB"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DecryptedWebhookInDB": {
                "title": "DecryptedWebhookInDB",
                "required": [
                    "id",
                    "token",
                    "channel_id",
                    "guild_id",
                    "created_at",
                    "updated_at",
                    "decrypted_token"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "token": {
                        "title": "Token",
                        "type": "string"
                    },
                    "channel_id": {
                        "title": "Channel Id",
                        "type": "string"
                    },
                    "guild_id": {
                        "title": "Guild Id",
                        "type": "string"
                    },
                    "is_nsfw": {
                        "title": "Is Nsfw",
                        "type": "boolean",
                        "default": true
                    },
                    "lang": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/WebhookLang"
                            }
                        ],
                        "default": "en"
                    },
                    "currency": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/WebhookCurrency"
                            }
                        ],
                        "default": "JPY"
                    },
                    "is_existed": {
                        "title": "Is Existed",
                        "type": "boolean",
                        "nullable": true
                    },
                    "created_at": {
                        "title": "Created At",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "type": "string",
                        "format": "date-time"
                    },
                    "decrypted_token": {
                        "title": "Decrypted Token",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
            "WebhookCreate": {
                "title": "WebhookCreate",
                "required": [
                    "id",
                    "token",
                    "channel_id",
                    "guild_id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "token": {
                        "title": "Token",
                        "type": "string"
                    },
                    "channel_id": {
                        "title": "Channel Id",
                        "type": "string"
                    },
                    "guild_id": {
                        "title": "Guild Id",
                        "type": "string"
                    },
                    "is_nsfw": {
                        "title": "Is Nsfw",
                        "type": "boolean",
                        "default": true
                    },
                    "lang": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/WebhookLang"
                            }
                        ],
                        "default": "en"
                    },
                    "currency": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/WebhookCurrency"
                            }
                        ],
                        "default": "JPY"
                    }
                }
            },
            "WebhookCurrency": {
                "title": "WebhookCurrency",
                "enum": [
                    "JPY",
                    "USD",
                    "EUR"
                ],
                "type": "string",
                "description": "An enumeration."
            },
            "WebhookLang": {
                "title": "WebhookLang",
                "enum": [
                    "zh-TW",
                    "en",
                    "ja"
                ],
                "type": "string",
                "description": "An enumeration."
            },
            "WebhookUpdate": {
                "title": "WebhookUpdate",
                "type": "object",
                "properties": {
                    "is_existed": {
                        "title": "Is Existed",
                        "type": "boolean",
                        "nullable": true
                    }
                }
            },
...
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-token"
            }
        }
    }
}